requirement.txt
flask
PyCryptodome
requests
